Add misc value r/w for trainer8
Closes #2639
casting object to generic T : struct isn't valid, so just cast the get/set appropriately

+++ PKHeX.Core/Saves/Substructures/Gen8/TrainerCard8.cs
@@ -31,7 +31,7 @@ namespace PKHeX.Core
                 var data = BitConverter.GetBytes(value);
                 SAV.SetData(Data, data, 0x28);
                 // set to the other block since it doesn't have an accessor
-                ((SAV8SWSH)SAV).SetValue(SaveBlockAccessorSWSH.KRotoRally, value);
+                ((SAV8SWSH)SAV).SetValue(SaveBlockAccessorSWSH.KRotoRally, (uint)value);
             }
         }
